+++ .prettierrc
@@ -1,6 +1,6 @@
 {
 	"tabWidth": 2,
-	"useTabs": true,
+	"useTabs": false,
 	"printWidth": 80,
 	"semi": true,
 	"trailingComma": "es5",
