+++ src/components/Button/Button.stories.js
@@ -2,18 +2,18 @@ import React from 'react';
 import Button from 'components/Button';
 
 export default {
-	title: 'Button',
-	component: Button,
+  title: 'Button',
+  component: Button,
 };
 
 const Template = (args) => {
-	return <Button {...args} />;
+  return <Button {...args} />;
 };
 
 const Default = Template.bind({});
 
 Default.args = {
-	children: <span>Default Button</span>,
+  children: <span>Default Button</span>,
 };
 
 export { Default };
