+++ src/core/App.js
@@ -1,55 +1,67 @@
-import React, { useEffect } from 'react';
+import React, { Suspense, useEffect } from 'react';
 import { Provider, useDispatch } from 'react-redux';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import { PersistGate } from 'redux-persist/integration/react';
 import { ToastContainer } from 'react-toastify';
 
-import { Home, Login } from 'pages';
-import PrivateRoute from 'core/PrivateRoute';
+// import PrivateRoute from 'core/PrivateRoute';
 import { store, persistor } from 'core/store';
 import { onAuthStateChanged } from 'firebase/auth';
 import { auth } from 'services/firebase';
-import { setUser } from 'global/redux/reducers/auth';
-import { PersistGate } from 'redux-persist/integration/react';
+import { setUser } from 'global/redux/auth/slice';
+
+import 'react-toastify/dist/ReactToastify.css';
+import routes from 'routes';
+import PrivateRoute from './PrivateRoute';
 
 function RoutesApp() {
-	const dispatch = useDispatch();
-	useEffect(() => {
-		const unsub = onAuthStateChanged(auth, (user) => {
-			dispatch(setUser(user));
-		});
+  const dispatch = useDispatch();
+  useEffect(() => {
+    const unsub = onAuthStateChanged(auth, (user) => {
+      dispatch(setUser(user));
+    });
 
-		return () => unsub();
-	}, []);
+    return () => unsub();
+  }, []);
 
-	return (
-		<Router>
-			<Routes>
-				<Route
-					path='/'
-					element={
-						<PrivateRoute>
-							<Home />
-						</PrivateRoute>
-					}
-				/>
-
-				<Route path='/login' element={<Login />} />
-
-				<Route path='*' element={<h1>Page not found</h1>} />
-			</Routes>
-		</Router>
-	);
+  return (
+    <Router>
+      <Suspense
+        fallback={<span style={{ backgroundColor: 'red' }}>Loading...</span>}
+      >
+        <Routes>
+          {routes.map((route) => (
+            <Route
+              key={route?.id}
+              path={route?.path}
+              element={
+                <PrivateRoute>
+                  <route.component />
+                </PrivateRoute>
+              }
+            />
+          ))}
+        </Routes>
+      </Suspense>
+    </Router>
+  );
 }
 
 function App() {
-	return (
-		<Provider store={store}>
-			<PersistGate loading={null} persistor={persistor}>
-				<ToastContainer />
-				<RoutesApp />
-			</PersistGate>
-		</Provider>
-	);
+  return (
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        <ToastContainer
+          autoClose={1000}
+          closeButton
+          position='top-right'
+          theme='light'
+          hideProgressBar
+        />
+        <RoutesApp />
+      </PersistGate>
+    </Provider>
+  );
 }
 
 export default App;
