+++ src/global/redux/auth/request.js
@@ -1,0 +1,23 @@
+import {
+  sendPasswordResetEmail,
+  signInWithEmailAndPassword,
+  signOut,
+} from 'firebase/auth';
+
+import { auth } from 'services/firebase';
+
+const authentication = {
+  signIn: (email, password) => {
+    signInWithEmailAndPassword(auth, email, password);
+  },
+
+  forgotPassword: async (email) => {
+    sendPasswordResetEmail(auth, email);
+  },
+
+  signOut: () => {
+    signOut(auth);
+  },
+};
+
+export default authentication;
