--- src/global/redux/reducers/auth.js
@@ -1,20 +1,0 @@
-const { createSlice } = require('@reduxjs/toolkit');
-
-const auth = createSlice({
-	name: 'authentication',
-
-	initialState: {
-		isLoading: false,
-		currentUser: { name: 'fas' },
-	},
-
-	reducers: {
-		setUser: (state, action) => {
-			state.currentUser = action.payload;
-		},
-	},
-});
-
-export const { setUser } = auth.actions;
-
-export default auth.reducer;
