--- src/pages/index.js
@@ -1,4 +1,0 @@
-import Home from 'pages/Home';
-import Login from 'pages/Login';
-
-export { Home, Login };
