--- src/utils/constants/toast.js
@@ -1,9 +1,0 @@
-const options = {
-	default: {
-		position: 'top-right',
-		autoClose: 1000,
-		closeOnClick: true,
-	},
-};
-
-export default options;
