+++ src/utils/helpers/index.js
@@ -1,0 +1,7 @@
+const accessToken = {
+  set: (token) => localStorage.setItem('accessToken', token),
+  get: () => localStorage.getItem('accessToken'),
+  remove: () => localStorage.removeItem('accessToken'),
+};
+
+export { accessToken };
