+++ src/views/PageNotFound/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const PageNotFound = () => {
+  return <div className='page-container'>Page not found</div>;
+};
+
+export default PageNotFound;
