+++ src/views/index.js
@@ -1,0 +1,5 @@
+import Home from 'views/Home';
+import Login from 'views/Login';
+import PageNotFound from 'views/PageNotFound';
+
+export { Home, Login, PageNotFound };
